+++ app/Http/Controllers/TransactionsController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\User;
 use Illuminate\Http\Request;
 use App\Product;
 use App\SingleTransaction;
@@ -14,8 +15,7 @@ class TransactionsController extends Controller
      *
      * @return void
      */
-    public function __construct()
-    {
+    public function __construct() {
         $this->middleware('auth');
     }
     
@@ -24,10 +24,13 @@ class TransactionsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
-        return view('pages.transactions')
-            ->with('transactions', Transaction::orderBy('created_at', 'desc')->paginate(15));
+    public function index() {
+        if ($this->isUserType('admin') || $this->isUserType('seller')) {
+            return view('pages.transactions')
+                ->with('transactions', Transaction::orderBy('created_at', 'desc')->paginate(20));
+        }
+
+        return redirect('/')->with('error', 'You don\'t have the privilege');
     }
 
     /**
@@ -35,10 +38,11 @@ class TransactionsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
-    {
-        return view('pages.transactions.create')
-            ->with('products', Product::all());
+    public function create() {
+        if ($this->isUserType('admin') || $this->isUserType('seller'))
+            return view('pages.transactions.create');
+
+        return redirect('/')->with('error', 'You don\'t have the privilege');
     }
 
     /**
@@ -47,48 +51,57 @@ class TransactionsController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
-        $capital = 0;
-        $income = 0;
+    public function store(Request $request) {
+        if ($this->isUserType('admin') || $this->isUserType('seller')) {
+            $capital = 0;
+            $income = 0;
 
-        $transaction = new Transaction;
-        $transaction->total = $request->get('total');
-        $transaction->money_received = $request->get('money');
-        $transaction->change = $request->get('change');
-        $transaction->capital = $capital;
-        $transaction->income = $income;
-        $transaction->save();
+            $transaction = new Transaction;
+            $transaction->total = $request->get('total');
+            $transaction->money_received = $request->get('money');
+            $transaction->change = $request->get('change');
+            $transaction->capital = $capital;
+            $transaction->income = $income;
+            $transaction->save();
 
-        foreach ($request->get('transactions') as $transac) {
-            $singleTransaction = new SingleTransaction;
-            $singleTransaction->transaction_id = $transaction->id;
-            $singleTransaction->product_id = $transac['product_id'];
-            $singleTransaction->quantity = $transac['quantity'];
-            $singleTransaction->total = $transac['subtotal'];
-            $singleTransaction->capital = $transac['quantity'] * $singleTransaction->product->price;
-            $singleTransaction->income = $transac['subtotal'] - $singleTransaction->capital;
-            $singleTransaction->orig_price = $singleTransaction->product->price;
-            $singleTransaction->orig_srp = $singleTransaction->product->srp;
-            $singleTransaction->save();
+            foreach ($request->get('transactions') as $transac) {
+                $singleTransaction = new SingleTransaction;
+                $singleTransaction->transaction_id = $transaction->id;
+                $singleTransaction->product_id = $transac['product_id'];
+                $singleTransaction->quantity = $transac['quantity'];
+                $singleTransaction->total = $transac['subtotal'];
+                $singleTransaction->capital = $transac['quantity'] * $singleTransaction->product->price;
+                $singleTransaction->income = $transac['subtotal'] - $singleTransaction->capital;
+                $singleTransaction->orig_price = $singleTransaction->product->price;
+                $singleTransaction->orig_srp = $singleTransaction->product->srp;
 
-            $product = Product::find($singleTransaction->product_id);
-            $product->stocks -= $singleTransaction->quantity;
-            $product->save();
+                $product = Product::find($singleTransaction->product_id);
+                $singleTransaction->name = $product->name;
+                $singleTransaction->type = $product->type;
+                $singleTransaction->desc = $product->desc;
+                $singleTransaction->save();
+                $product->stocks -= $singleTransaction->quantity;
+                $product->save();
 
-            $capital += $singleTransaction->capital;
-            $income += $singleTransaction->income;
+                $capital += $singleTransaction->capital;
+                $income += $singleTransaction->income;
+            }
+
+            $transaction = Transaction::find($transaction->id);
+            $transaction->capital = $capital;
+            $transaction->income = $income;
+            $transaction->save();
+
+            $data = array(
+                'text' => "success",
+                'id' => $transaction->id
+            );
+
+            return json_encode($data);
         }
 
-        $transaction = Transaction::find($transaction->id);
-        $transaction->capital = $capital;
-        $transaction->income = $income;
-        $transaction->save();
 
-        return "success";
-
-        // $returnHTML = view('pages.transactions')->with('success', 'Transaction Finished')->render();
-        // return response()->json(array('success' => true, 'html'=>$returnHTML));
+        return redirect('/')->with('error', 'You don\'t have the privilege');
     }
 
     /**
@@ -97,10 +110,11 @@ class TransactionsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
-    {
-        return view('pages.transactions.single')
-            ->with('transaction', Transaction::find($id));
+    public function show($id) {
+        if ($this->isUserType('admin') || $this->isUserType('seller'))
+            return view('pages.transactions.single')->with('transaction', Transaction::find($id));
+
+        return redirect('/')->with('error', 'You don\'t have the privilege');
     }
 
     /**
@@ -109,8 +123,7 @@ class TransactionsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
-    {
+    public function edit($id) {
         //
     }
 
@@ -121,8 +134,7 @@ class TransactionsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
-    {
+    public function update(Request $request, $id) {
         //
     }
 
@@ -132,27 +144,30 @@ class TransactionsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
-    {
-        $transaction = Transaction::find($id);
-        $transaction->delete();
+    public function destroy($id) {
+        if ($this->isUserType('admin') || $this->isUserType('seller')){
+            $transaction = Transaction::find($id);
+            $transaction->delete();
 
-        return redirect('/transactions')->with('success', 'Transaction Deleted');
+            return redirect('/transactions')->with('success', 'Transaction Deleted');
+        }
+
+        return redirect('/')->with('error', 'You don\'t have the privilege');
     }
 
     public function get(Request $request) {
         if ($request->ajax()) return $this->store($request);
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function success()
-    {
-        return view('pages.transactions')
-            ->with('success', 'Added Transaction Successfully')
-            ->with('transactions', Transaction::orderBy('created_at', 'desc')->paginate(15));
+    public function success() {
+        if ($this->isUserType('admin') || $this->isUserType('seller')){
+            return view('pages.transactions')
+                ->with('success', 'Added Transaction Successfully')
+                ->with('transactions', Transaction::orderBy('created_at', 'desc')->paginate(15));
+        }
+
+        return redirect('/')->with('error', 'You don\'t have the privilege');
     }
+
+    public function isUserType($type) { return (User::find(auth()->user()->id)->type == $type) ? true : false; }
 }
